Refactored the cli and fixed another Heisenberg bug
