+++ src/pepsflow/models/tensors.py
@@ -87,10 +87,8 @@ class Tensors:
         """
         rot = torch.matrix_exp(-1j * torch.pi / 2 * sy)
         sz = torch.complex(sz, torch.zeros_like(sz))
-        res = (
-            lam * torch.kron(sz, sz) / 4
-            + torch.kron(sp, sm) / 2
-            + torch.kron(sm, sp) / 2
+        res = lam * (
+            torch.kron(sz, sz) / 4 + torch.kron(sp, sm) / 2 + torch.kron(sm, sp) / 2
         )
         res = res.view(2, 2, 2, 2)
         res = torch.einsum("abcd,be,df->aefc", res, rot, torch.conj(rot)).real
