+++ src/pepsflow/optimize.cfg
@@ -1,15 +1,15 @@
 [PARAMETERS]
-model = Heisenberg
-chi = 32
+model = Ising
+chi = 64
 d = 2
-lam = [2.0, 2.1, 2.2, 2.3, 2.4, 2.5, 2.6, 2.7, 2.8, 2.9, 3.0]
+lam = [2.75]
 max_iter = 10
 runs = 5
 learning_rate = 1.0
-epochs = 20
+epochs = 10
 perturbation = 0.0
 gpu = False
-threads = 2
-save_folder = test
+threads = 1
+save_folder = d=2
 data_folder = None
